requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
requests>=2.28.0
